Added Equations that were used
I could have put all the values for every trigonometric function into the C++ program, but that would not be logical because the values for one function can solve for all of the others.  In this case, I used cosine.  Therefore, the values of cosine are inserted in the Cos() function, while the other trigonometric functions without inverses use the Cos() function to solve.  In the uploaded picture are the equations that were used to do this.